% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.dots_err}
\alias{.dots_err}
\title{Dots Error Condition}
\usage{
.dots_err(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[base]{simpleError}()}}
}
\value{
An error condition with the message
\dQuote{\code{'...' must be empty}}
}
\description{
Generate an error condition for incorrect dots (\dots) usage
}
\examples{
.dots_err()

}
\seealso{
Other internal utility functions:
\code{\link{.info_err}()},
\code{\link{.is_na}()},
\code{\link{.is_string}()},
\code{\link{.is_var}()}
}
\concept{utils}
\keyword{error}
\keyword{internal}
\keyword{utilities}
