% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.info_err}
\alias{.info_err}
\title{Info Error Message}
\usage{
.info_err(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[base]{simpleError}()}}
}
\value{
An error condition with the message
\dQuote{\code{'info' must be a single character value}}
}
\description{
Generate an error condition for incorrect \sQuote{\code{info}} usage
}
\examples{
.info_err()

}
\seealso{
Other internal utility functions:
\code{\link{.dots_err}()},
\code{\link{.is_na}()},
\code{\link{.is_string}()},
\code{\link{.is_var}()}
}
\concept{utils}
\keyword{error}
\keyword{internal}
\keyword{utilities}
