% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.is_na}
\alias{.is_na}
\title{Is an Object \code{NA}}
\usage{
.is_na(x)
}
\arguments{
\item{x}{An \R object}
}
\value{
\code{TRUE} if \code{x} is \code{NA}, otherwise \code{FALSE}
}
\description{
Test that an object is \code{NA}; an object is \code{NA} if
\itemize{
 \item it has a length of \eqn{1}
 \item \code{is.na(x)} is \code{TRUE}
}
}
\examples{
.is_na(NA) # TRUE
.is_na(c(NA, NA)) # FALSE

}
\seealso{
Other internal utility functions:
\code{\link{.dots_err}()},
\code{\link{.info_err}()},
\code{\link{.is_string}()},
\code{\link{.is_var}()}
}
\concept{utils}
\keyword{NA}
\keyword{internal}
\keyword{utilities}
