% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.is_string}
\alias{.is_string}
\title{Is an Object a String}
\usage{
.is_string(x)
}
\arguments{
\item{x}{An \R object}
}
\value{
\code{TRUE} if \code{x} is a string, otherwise \code{FALSE}
}
\description{
Test that an object is a string; an object is a string if
\itemize{
 \item it is a \link[base]{character}
 \item it has a length of \eqn{1}
 \item is is not \code{NA}
 \item it is not \link[base:nzchar]{empty}
}
}
\examples{
.is_string("a") # TRUE
.is_string("hello") # TRUE
.is_string("hello world") # FALSE

.is_string("") # FALSE
.is_string(character(length = 0L)) # FALSE
.is_string(c("hello", "world")) # FALSE

}
\seealso{
\code{\link[base]{nzchar}()}

Other internal utility functions:
\code{\link{.dots_err}()},
\code{\link{.info_err}()},
\code{\link{.is_na}()},
\code{\link{.is_var}()}
}
\concept{utils}
\keyword{character}
\keyword{internal}
\keyword{utilities}
