% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.is_var}
\alias{.is_var}
\title{Is an Environment Variable Trueish}
\usage{
.is_var(var)
}
\arguments{
\item{var}{An environment variable}
}
\value{
\code{TRUE} if \code{var} is trueish, otherwise \code{FALSE}
}
\description{
Is an Environment Variable Trueish
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
withr::with_envvar(c(R_TINYEXPECT_ENVVAR = "true"), {
  .is_var("R_TINYEXPECT_ENVVAR") # TRUE
})

withr::with_envvar(c(R_TINYEXPECT_ENVVAR = "false"), {
  .is_var("R_TINYEXPECT_ENVVAR") # FALSE
})
\dontshow{\}) # examplesIf}
}
\seealso{
Other internal utility functions:
\code{\link{.dots_err}()},
\code{\link{.info_err}()},
\code{\link{.is_na}()},
\code{\link{.is_string}()}
}
\concept{utils}
\keyword{internal}
\keyword{utilities}
