% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_compare.R
\name{expect_compare}
\alias{expect_compare}
\title{Expect a Comparison to Hold}
\usage{
expect_compare(
  current,
  target,
  ...,
  operator,
  info = NA_character_,
  call = NULL
)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{target}{\code{[R object or expression]} Expected outcome}

\item{\dots}{Ignored}

\item{operator}{\code{[character(1L)]} Name of comparison function for
expectation; choose from:
\itemize{
 \item \dQuote{\code{<}} for strictly less than
 \item \dQuote{\code{<=}} for less than or equal to
 \item \dQuote{\code{>}} strictly greater than
 \item \dQuote{\code{>=}} for greater than or equal to
 \item \dQuote{\code{==}} for equal to
 \item \dQuote{\code{!=}} for not equal to
}}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}

\item{call}{\code{[call]} An optional \code{\link[base]{call}} object for
for expectation predicates that wrap \code{expect_compare()}}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing based
on \link[base:Comparison]{comparisons}
}
\examples{
# Expect strictly less than
expect_compare(1L, 3L, operator = "<") # Pass
expect_compare(3L, 1L, operator = "<") # Fail
expect_compare(1L, 1L, operator = "<") # Fail

# Expect less than or equal to
expect_compare(1L, 3L, operator = "<=") # Pass
expect_compare(3L, 1L, operator = "<=") # Fail
expect_compare(1L, 1L, operator = "<=") # Pass

# Expect strictly greater than
expect_compare(1L, 3L, operator = ">") # Fail
expect_compare(3L, 1L, operator = ">") # Pass
expect_compare(1L, 1L, operator = ">") # Fail

# Expect greater than or equal to
expect_compare(1L, 3L, operator = ">=") # Fail
expect_compare(3L, 1L, operator = ">=") # Pass
expect_compare(1L, 1L, operator = ">=") # Pass

# Expect equal to
expect_compare(1L, 3L, operator = "==") # Fail
expect_compare(3L, 1L, operator = "==") # Fail
expect_compare(1L, 1L, operator = "==") # Pass

# Expect not equal to
expect_compare(1L, 3L, operator = "!=") # Pass
expect_compare(3L, 1L, operator = "!=") # Pass
expect_compare(1L, 1L, operator = "!=") # Fail

}
\seealso{
\link[base:Comparison]{Comparison operators} in \R

Convenience functions: \code{\link{expect_lt}()},
\code{\link{expect_lte}()}, \code{\link{expect_gt}()}, and
\code{\link{expect_gte}()}
}
\keyword{internal}
