% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_comparisons.R
\name{expect_comparisons}
\alias{expect_comparisons}
\alias{expect_lt}
\alias{expect_lte}
\alias{expect_gt}
\alias{expect_gte}
\title{Expect a Comparison to Hold}
\usage{
expect_lt(current, target, ..., info = NA_character_)

expect_lte(current, target, ..., info = NA_character_)

expect_gt(current, target, ..., info = NA_character_)

expect_gte(current, target, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{target}{\code{[R object or expression]} Expected outcome}

\item{...}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing based
on \link[base:Comparison]{comparisons}
}
\examples{
# Expect strictly less than
expect_lt(1L, 3L) # Pass
expect_lt(3L, 1L) # Fail
expect_lt(1L, 1L) # Fail

# Expect less than or equal to
expect_lte(1L, 3L) # Pass
expect_lte(3L, 1L) # Fail
expect_lte(1L, 1L) # Pass

# Expect strictly greater than
expect_gt(1L, 3L) # Fail
expect_gt(3L, 1L) # Pass
expect_gt(1L, 1L) # Fail

# Expect greater than or equal to
expect_gte(1L, 3L) # Fail
expect_gte(3L, 1L) # Pass
expect_gte(1L, 1L) # Pass

}
\seealso{
\CRANpkg{testthat} equivalents:
\code{\link[testthat:comparison-expectations]{expect_lt}()},
\code{\link[testthat:comparison-expectations]{expect_lte}()},
\code{\link[testthat:comparison-expectations]{expect_gt}()}, and
\code{\link[testthat:comparison-expectations]{expect_gte}()}
}
