% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_condition.R
\name{expect_condition}
\alias{expect_condition}
\title{Expect a Condition to be Thrown}
\usage{
expect_condition(
  current,
  pattern = NULL,
  class = NULL,
  ...,
  info = NA_character_
)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{pattern}{\code{[character(1L)]} A regular expression to match the
condition message; any condition thrown that does not match \code{pattern}
will trigger a failure. Must be a single character string}

\item{class}{\code{[character]} Optional vector of classes from which the
condition should inherit; any conditions thrown that do no inherit from
\code{class} will trigger a failure}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that a given function call throws a
\link[base]{condition}
}
\examples{
expect_condition(message("test message")) # Pass
expect_condition(warning("test warning")) # Pass
expect_condition(stop("test error")) # Pass

expect_condition("tomato") # Fail
expect_condition(message("test message"), class = "specialMessage") # Fail
expect_condition(message("test message"), pattern = "tomato") # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_condition]{testthat::expect_condition}()}

Other condition expectation predicates:
\code{\link{expect_no_condition}()}
}
\concept{condition}
