% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_contains.R
\name{expect_contains}
\alias{expect_contains}
\title{Expect a Superset}
\usage{
expect_contains(current, target, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R vector]} A \link[base]{vector} to be tested}

\item{target}{\code{[R vector]} A \link[base]{vector} to be compared against}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that a given vector contains all values
present in another vector
}
\examples{
expect_contains(letters, target = rev(letters)) # Pass
expect_contains(letters, target = "b") # Pass
expect_contains("b", target = letters) # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_contains]{testthat::expect_contains}()}

Other in-condition expectation predicates:
\code{\link{expect_in}()},
\code{\link{expect_mapequal}()},
\code{\link{expect_setequal}()}
}
\concept{set}
