% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_in.R
\name{expect_in}
\alias{expect_in}
\title{Expect a Subset}
\usage{
expect_in(current, target, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R vector]} A \link[base]{vector} to be tested}

\item{target}{\code{[R vector]} A \link[base]{vector} to be compared against}

\item{...}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that a vector is fully contained within
another vector
}
\examples{
expect_in(letters, target = rev(letters)) # Pass
expect_in(letters, target = "b") # Fail
expect_in("b", target = letters) # Pass

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_in]{testthat::expect_in}()}

Other in-condition expectation predicates:
\code{\link{expect_contains}()},
\code{\link{expect_mapequal}()},
\code{\link{expect_setequal}()}
}
\concept{set}
