% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_mapequal.R
\name{expect_mapequal}
\alias{expect_mapequal}
\title{Expect Map Equality}
\usage{
expect_mapequal(
  current,
  target,
  ...,
  info = NA_character_,
  tolerance = sqrt(x = .Machine$double.eps)
)
}
\arguments{
\item{current}{\code{[R vector]} A \link[base]{vector} to be tested}

\item{target}{\code{[R vector]} A \link[base]{vector} to be compared against}

\item{...}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}

\item{tolerance}{\code{[numeric]} Test equality to machine rounding; passed
to \code{\link[base]{all.equal}()}}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that the named \link[base]{sets} of
two vectors are equal
}
\examples{
(x <- `names<-`(letters, letters))
expect_mapequal(x, target = rev(x)) # Pass
expect_mapequal(letters, target = c(b = "b")) # Fail
expect_mapequal(c(b = "b"), target = x) # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_mapequal]{testthat::expect_mapequal}()}

Other in-condition expectation predicates:
\code{\link{expect_contains}()},
\code{\link{expect_in}()},
\code{\link{expect_setequal}()}
}
\concept{set}
