% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_named.R
\name{expect_named}
\alias{expect_named}
\title{Expect Names}
\usage{
expect_named(
  current,
  target,
  ...,
  ignore.order = FALSE,
  ignore.case = FALSE,
  info = NA_character_
)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{target}{\code{[missing, character, or NULL]} One of the following:
\itemize{
 \item Left \link[base]{missing} to test that \code{current} has names
 \item A character vector giving the expected names of \code{current}
 \item \code{NULL} to test that \code{current} is unnamed
}}

\item{\dots}{Ignored}

\item{ignore.order}{\code{[logical(1L)]} Ignore the order of
\code{names(current)} with respect to \code{target}}

\item{ignore.case}{\code{[logical(1L)]} Ignore upper- vs lower-case
discrepancies between \code{names(current)} and \code{target}}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing the \link[base]{names} of an object
}
\examples{
(x <- c(a = 1L, b = 2L, c = 3L))
(y <- c(4L, 5L, 6L))

expect_named(x) # Pass
expect_named(y) # Fail

expect_named(x, target = NULL) # Fail
expect_named(y, target = NULL) # Pass

expect_named(x, target = c("a", "b", "c")) # Pass
expect_named(y, target = c("a", "b", "c")) # Fail

expect_named(x, target = c("b", "a", "c")) # Fail
expect_named(x, target = c("b", "a", "c"), ignore.order = TRUE) # Pass

expect_named(x, target = c("A", "B", "C")) # Fail
expect_named(x, target = c("A", "B", "C"), ignore.case = TRUE) # Pass

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_named]{testthat::expect_named}()}
}
