% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_no_condition.R
\name{expect_no_condition}
\alias{expect_no_condition}
\title{Expect No Conditions to be Thrown}
\usage{
expect_no_condition(
  current,
  pattern = NULL,
  class = NULL,
  ...,
  info = NA_character_
)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{pattern}{\code{[character(1L)]} A regular expression to match the
condition message; any conditions thrown with a message matching
\code{pattern} will trigger a failure while all other conditions
thrown will trigger a success. Must be a single character string}

\item{class}{\code{[character]} Optional vector of condition classes that
should \strong{not} be thrown; conditions thrown that inherit from
\code{class} will trigger a failure while all other conditions thrown
will trigger a success. If \code{NULL}, any condition thrown will
trigger a failure}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that a given function does not raise
a \link[base]{condition}
}
\examples{
expect_no_condition("tomato") # Pass

expect_no_condition(message("test")) # Fail

expect_no_condition(message("test"), pattern = "test") # Fail
expect_no_condition(message("test"), pattern = "tomato") # Pass

expect_no_condition(message("test"), class = "message") # Fail
expect_no_condition(message("test"), class = "error") # Pass

expect_no_condition(message("test"), pattern = "test", class = "message") # Fail
expect_no_condition(message("test"), pattern = "tomato", class = "message") # Pass
expect_no_condition(message("test"), pattern = "test", class = "error") # Pass
expect_no_condition(message("test"), pattern = "tomato", class = "error") # Pass

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_no_condition]{testthat::expect_no_condition}()}

Other condition expectation predicates:
\code{\link{expect_condition}()}
}
\concept{condition}
