% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_s3_class.R
\name{expect_s3_class}
\alias{expect_s3_class}
\title{Expect S3 Object}
\usage{
expect_s3_class(current, class, ..., info = NA_character_, exact = FALSE)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{class}{\code{[character]} Expected class of \code{current}; may
optionally pass \code{NA} to assert that \code{current} is \strong{not}
an object}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}

\item{exact}{\code{[logical(1L)]} If \code{TRUE}, check that
\code{class(current)} is identical to \code{class}; otherwise,
check that \code{current} \link[base:inherits]{inherits} \code{class}}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Check that \code{current} is an S3 object that
\link[base:inherits]{inherits} from \code{class}
}
\examples{
expect_s3_class(data.frame(), "data.frame") # Pass
expect_s3_class(1L, "integer") # Fail
expect_s3_class(1L, NA) # Pass

\dontshow{if (requireNamespace("Matrix", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
expect_s3_class(Matrix::Matrix(), "Matrix") # Fail
expect_s3_class(Matrix::Matrix(), NA) # Pass
\dontshow{\}) # examplesIf}
}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_s3_class]{testthat::expect_s3_class}()}

Other typing expectation predicates:
\code{\link{expect_s4_class}()},
\code{\link{expect_type}()}
}
\concept{inheritance}
