% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_s4_class.R
\name{expect_s4_class}
\alias{expect_s4_class}
\title{Expect S4 Object}
\usage{
expect_s4_class(current, class, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{class}{\code{[character]} Expected class of \code{current}; may
optionally pass \code{NA} to assert that \code{current} is \strong{not}
an object}

\item{...}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single character string}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Check that \code{current} is an S4 object that is of class \code{class}
}
\examples{
if (requireNamespace("Matrix", quietly = TRUE)) {
  expect_s4_class(Matrix::Matrix(), "Matrix")
}

expect_s4_class(data.frame(), "data.frame")
expect_s4_class(data.frame(), NA)

}
\seealso{
\pkg{testthat} equivalent:
\code{\link[testthat:expect_s4_class]{testthat::expect_s4_class}()}

Other inheritance: 
\code{\link{expect_s3_class}()},
\code{\link{expect_type}()}
}
\concept{inheritance}
