% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_setequal.R
\name{expect_setequal}
\alias{expect_setequal}
\title{Expect Set Equality}
\usage{
expect_setequal(current, target, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R vector]} A \link[base]{vector} to be tested}

\item{target}{\code{[R vector]} A \link[base]{vector} to be compared against}

\item{...}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expectation predicate for testing that the \link[base]{sets} of two vectors
are equal
}
\examples{
expect_setequal(letters, target = rev(letters)) # Pass
expect_setequal(letters, target = "b") # Fail
expect_setequal("b", target = letters) # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_setequal]{testthat::expect_setequal}()}

Other in-condition expectation predicates:
\code{\link{expect_contains}()},
\code{\link{expect_in}()},
\code{\link{expect_mapequal}()}
}
\concept{set}
