% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_type.R
\name{expect_type}
\alias{expect_type}
\title{Expect Object Type}
\usage{
expect_type(current, type, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{type}{\code{[character(1L)]} Expected base type (as given by
\code{\link{typeof}()})}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Check that \code{typeof(current)} is \code{type}
}
\examples{
expect_type(1L, "integer") # Pass
expect_type(1.0, "integer") # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_type]{testthat::expect_type}()}

Other typing expectation predicates:
\code{\link{expect_s3_class}()},
\code{\link{expect_s4_class}()}
}
\concept{inheritance}
