% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_visible.R
\name{expect_visible}
\alias{expect_visible}
\title{Expect Visible}
\usage{
expect_visible(current, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single character string}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expect Visible
}
\examples{
f <- \() invisible()
expect_visible(f())

g <- \() NULL
expect_visible(g())


}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_visible]{testthat::expect_visible}()}

\code{\link{withVisible}()}

Other visibility: 
\code{\link{expect_invisible}()}
}
\concept{visibility}
