% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\title{Stop Testing}
\usage{
skip(message = "Skipping")
}
\arguments{
\item{message}{A message describing why the test file is being skipped}
}
\value{
If called within a \CRANpkg{tinytest} test, triggers an exit
condition; otherwise, returns \code{message}
}
\description{
Unconditionally stop testing a \CRANpkg{tinytest} test file, preventing
additional tests in the file from running without triggering a failure. This
is the low-level exit function for other \code{skip_()} functions in
\pkg{tinyexpect}
}
\examples{
skip()

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:skip]{testthat::skip}()}

\code{\link[tinytest:exit_file]{tinytest::exit_file}()}

Other "stop testing" functions:
\code{\link{skip_if_not_installed}()},
\code{\link{skip_on_bioc}()},
\code{\link{skip_on_ci}()},
\code{\link{skip_on_covr}()},
\code{\link{skip_on_cran}()},
\code{\link{skip_on_os}()}
}
\concept{skip}
