% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_if_not_installed.R
\name{skip_if_not_installed}
\alias{skip_if_not_installed}
\alias{exit_if_not_installed}
\title{Stop Testing if a Required Package is Not Installed}
\usage{
skip_if_not_installed(pkg, minimum_version = NULL, quietly = TRUE)

exit_if_not_installed(pkg, minimum_version = NULL, quietly = TRUE)
}
\arguments{
\item{pkg}{Name of package to check for}

\item{minimum_version}{Optional minimum required version of \code{pkg}}

\item{quietly}{Attempt to find the package quietly; passed to
\code{\link[base]{requireNamespace}()}}
}
\value{
If called within a \CRANpkg{tinytest} test and \code{pkg} is either
not installed or not at least \code{minimum_version}, triggers an exit
condition; otherwise, returns one of
\itemize{
 \item A string saying that \code{pkg} is not installed
 \item A string saying that \code{pkg} is installed, but not
  at least \code{minimum_version}
 \item \code{NULL} invisibly
}
}
\description{
Conditionally stop testing a \CRANpkg{tinytest} test file if a required
package is not available or not of a minimum required version
}
\section{\dQuote{Skip} vs \dQuote{Exit}}{

\pkg{tinyexpect} provides both \dQuote{\code{skip_}} and \dQuote{\code{exit_}}
versions of \dQuote{stop testing} functions due to the different philosophies
of \CRANpkg{tinytest} and \CRANpkg{testthat}; in \CRANpkg{testthat}, tests
are encapsulated by \code{\link[testthat]{test_that}()} to create smaller
testing units within a single test file. As such, if a series of tests need
to be passed over for some reason, it makes sense to \dQuote{skip} a
\code{\link[testthat]{test_that}()} block and move on to the next block

\CRANpkg{tinytest}, however, treats each test file as a testing unit. Each
file in \code{inst/tinytest} is equivalent to a \CRANpkg{testthat}
\code{\link[testthat]{test_that}()} block; as such, if a series of tests need
to be passed over for some reason, it makes sense to \dQuote{exit} a test
file and move on to the next file in \code{inst/tinytest}

In order to provide compatibility with users transitioning from
\CRANpkg{testthat} to \CRANpkg{tinytest}, and to provide continuity with the
\CRANpkg{tinytest} philosophy, \pkg{tinyexpect} provides both \code{skip_}-
and \code{exit_}- prefixed \dQuote{stop testing} functions that work
identically to one another
}

\examples{
(pkg <- paste(sample(letters, size = 7L, replace = TRUE), collapse = ""))
skip_if_not_installed(pkg)

skip_if_not_installed("tinyexpect", minimum_version = "99.0.1")

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:skip_if_not_installed]{testthat::skip_if_not_installed}()}

Other "stop testing" functions:
\code{\link{skip}()},
\code{\link{skip_on_bioc}()},
\code{\link{skip_on_ci}()},
\code{\link{skip_on_covr}()},
\code{\link{skip_on_cran}()},
\code{\link{skip_on_os}()}
}
\concept{skip}
