% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_on_ci.R
\name{skip_on_ci}
\alias{skip_on_ci}
\alias{exit_on_ci}
\title{Stop Testing on CI}
\usage{
skip_on_ci()

exit_on_ci()
}
\value{
If called within a \CRANpkg{tinytest} test running on CI, triggers
an exit condition; otherwise, either a string saying \dQuote{\code{On CI}}
or \code{NULL} invisibly
}
\description{
Conditionally stop testing a \CRANpkg{tinytest} test file if running on CI.
This is determined by checking if the environment variable \dQuote{\code{CI}}
set and is a non-\code{FALSE} value
}
\section{\dQuote{Skip} vs \dQuote{Exit}}{

\pkg{tinyexpect} provides both \dQuote{\code{skip_}} and \dQuote{\code{exit_}}
versions of \dQuote{stop testing} functions due to the different philosophies
of \CRANpkg{tinytest} and \CRANpkg{testthat}; in \CRANpkg{testthat}, tests
are encapsulated by \code{\link[testthat]{test_that}()} to create smaller
testing units within a single test file. As such, if a series of tests need
to be passed over for some reason, it makes sense to \dQuote{skip} a
\code{\link[testthat]{test_that}()} block and move on to the next block

\CRANpkg{tinytest}, however, treats each test file as a testing unit. Each
file in \code{inst/tinytest} is equivalent to a \CRANpkg{testthat}
\code{\link[testthat]{test_that}()} block; as such, if a series of tests need
to be passed over for some reason, it makes sense to \dQuote{exit} a test
file and move on to the next file in \code{inst/tinytest}

In order to provide compatibility with users transitioning from
\CRANpkg{testthat} to \CRANpkg{tinytest}, and to provide continuity with the
\CRANpkg{tinytest} philosophy, \pkg{tinyexpect} provides both \code{skip_}-
and \code{exit_}- prefixed \dQuote{stop testing} functions that work
identically to one another
}

\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
withr::with_envvar(c(CI = "true"), skip_on_ci())
\dontshow{\}) # examplesIf}
}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:skip_on_ci]{testthat::skip_on_ci}()}

Other "stop testing" functions:
\code{\link{skip}()},
\code{\link{skip_if_not_installed}()},
\code{\link{skip_on_bioc}()},
\code{\link{skip_on_covr}()},
\code{\link{skip_on_cran}()},
\code{\link{skip_on_os}()}
}
\concept{skip}
