% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_invisible.R
\name{expect_invisible}
\alias{expect_invisible}
\title{Expect Invisible}
\usage{
expect_invisible(current, ..., info = NA_character_)
}
\arguments{
\item{current}{\code{[R object or expression]} Object or expression to be tested}

\item{\dots}{Ignored}

\item{info}{\code{[character(1L)]} Optional user-defined message; must be a
single string value}
}
\value{
A \code{\link[tinytest]{tinytest}} object. A tinytest object is a
\code{logical} with attributes holding information about the
test that was run
}
\description{
Expect Invisible
}
\examples{
f <- function() invisible()
expect_invisible(f()) # Pass

g <- function() NULL
expect_invisible(g()) # Fail

}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:expect_invisible]{testthat::expect_invisible}()}

\code{\link{withVisible}()}, \code{\link{invisible}()}

Other visiblity expectation predicate:
\code{\link{expect_visible}()}
}
\concept{visibility}
