% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_on_bioc.R
\name{skip_on_bioc}
\alias{skip_on_bioc}
\alias{exit_on_bioc}
\title{Stop Testing on the Bioconductor Build System}
\usage{
skip_on_bioc()

exit_on_bioc()
}
\value{
If called within a \CRANpkg{tinytest} test running on the
Bioconductor Build System, triggers an exit condition; otherwise, either
a string saying \dQuote{\code{On Bioconductor}} or \code{NULL} invisibly
}
\description{
Conditionally stop testing a \CRANpkg{tinytest} test file if running on
the Bioconductor Build System. This is determined by checking if the
environment variable \dQuote{\code{IS_BIOC_BUILD_MACHINE}} set and is
a non-\code{FALSE} value
}
\examples{
\dontshow{if (requireNamespace("withr", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
withr::with_envvar(c(IS_BIOC_BUILD_MACHINE = "true"), skip_on_bioc())
\dontshow{\}) # examplesIf}
}
\seealso{
\CRANpkg{testthat} equivalent:
\code{\link[testthat:skip_on_bioc]{testthat::skip_on_bioc}()}

Other "stop testing" functions:
\code{\link{skip}()},
\code{\link{skip_if_not_installed}()},
\code{\link{skip_on_ci}()},
\code{\link{skip_on_covr}()},
\code{\link{skip_on_cran}()},
\code{\link{skip_on_os}()}
}
\concept{skip}
