% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tinyexpect-package}
\alias{tinyexpect}
\alias{tinyexpect-package}
\title{tinyexpect: Additional Expecation Predicates for the 'tinytest' Testing Framework}
\description{
Ports of 'testthat' expectation predicates to the 'tinytest' testing framework. These ports allow easier transitioning from 'testthat' to 'tinytest' as well as provide more explicit expectation predicates for 'tinytest' users.
}
\section{\CRANpkg{testthat} Equivalents}{

\pkg{tinyexpect} aims to port as many \CRANpkg{testthat} expectation
predicates to the \CRANpkg{tinytest} framework as makes sense.
\Sexpr[stage=render,results=rd]{
eq <- Filter(
  f = function(x) startsWith(x, "expect_"),
  x = tinyexpect:::.equivs$testthat$equivalents
)
if (length(eq)) {
  msg <- paste(
    "The following",
    ngettext(length(eq), "predicate has", "predicates have"),
    "been ported:",
    sprintf(
      "\\\itemize{\n\%s\n}",
      paste0(
        sprintf("\\\item \\\code{\\\\link{\%s}()}", eq),
        collapse = "\n"
      )
    )
  )
  missing <- tinyexpect:::.equivs$testthat$missing
  if (length(missing)) {
    mm <- sprintf(
      paste(
        "\\\subsection{Missing Predicates}{The following \\\pkg{\%s}",
        "\%s not been ported:\n\%s}"
      ),
      "testthat",
      ngettext(length(missing), "predicate has", "predicates have"),
      sprintf(
        "\\\itemize{\n\%s\n}",
        paste0(
          sprintf("\\\item \\\code{\\\\link[testthat]{\%s}()}", missing),
          collapse = "\n"
        )
      )
    )
    msg <- paste(msg, mm)
  }
  utilities <- Filter(
    f = function(x) !startsWith(x, "expect_"),
    x = tinyexpect:::.equivs$testthat$equivalents
  )
  if (length(utilities)) {
    mu <- sprintf(
      paste(
        "\\\subsection{Utility Functions}{The following \\\pkg{\%s}",
        "utility \%s been ported:\n\%s}"
      ),
      "testthat",
      ngettext(length(utilities), "function has", "functions have"),
      sprintf(
        "\\\itemize{\n\%s\n}",
        paste0(
          sprintf("\\\item \\\code{\\\\link{\%s}()}", utilities),
          collapse = "\n"
        )
      )
    )
    msg <- paste(msg, mu)
  }
  msg
} else {
  paste(
    "For a complete list of equivalent predicates,",
    "please install and load \\\pkg{testthat}"
  )
}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mojaveazure.github.io/tinyexpect}
  \item \url{https://github.com/mojaveazure/tinyexpect}
  \item Report bugs at \url{https://github.com/mojaveazure/tinyexpect/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{hoff0792@alumni.umn.edu} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

}
\keyword{internal}
\keyword{package}
